CORE-8123 Move logic from DC dialog and view into presenter
