+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/gin/factory/DeployedComponentListingViewFactory.java
@@ -4,12 +4,14 @@ import org.iplantc.de.apps.integration.client.view.tools.DeployedComponentsListi
 import org.iplantc.de.client.models.tool.Tool;
 
 import com.sencha.gxt.data.shared.ListStore;
-import com.sencha.gxt.widget.core.client.selection.SelectionChangedEvent;
+import com.sencha.gxt.data.shared.loader.FilterPagingLoadConfig;
+import com.sencha.gxt.data.shared.loader.PagingLoadResult;
+import com.sencha.gxt.data.shared.loader.PagingLoader;
 
 /**
  * @author jstroot
  */
 public interface DeployedComponentListingViewFactory {
     DeployedComponentsListingView createDcListingView(ListStore<Tool> listStore,
-                                                      SelectionChangedEvent.SelectionChangedHandler<Tool> handler);
+                                                      PagingLoader<FilterPagingLoadConfig, PagingLoadResult<Tool>> loader);
 }
