+++ de-lib/src/main/java/org/iplantc/de/apps/integration/client/view/tools/DeployedComponentsListingView.java
@@ -9,14 +9,18 @@ import com.google.gwt.resources.client.ImageResource;
 import com.google.gwt.safehtml.shared.SafeHtml;
 import com.google.gwt.user.client.ui.IsWidget;
 
+import com.sencha.gxt.widget.core.client.selection.SelectionChangedEvent;
+
 import java.util.List;
 
 /**
  * @author sriram
  *
  */
-public interface DeployedComponentsListingView extends IsWidget {
-    public interface Presenter extends org.iplantc.de.commons.client.presenter.Presenter {
+public interface DeployedComponentsListingView extends IsWidget,
+                                                       SelectionChangedEvent.HasSelectionChangedHandlers<Tool> {
+    public interface Presenter extends org.iplantc.de.commons.client.presenter.Presenter,
+                                       SelectionChangedEvent.HasSelectionChangedHandlers<Tool> {
 
         Tool getSelectedDC();
     }
